# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Pool API documentation",
        "contact": {},
        "version": "3.0.0"
    },
    "host": "pool3-api.vivaconagua.org",
    "basePath": "/v1",
    "paths": {
...
        "/mails/mailbox/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/mailbox"
                ],
                "summary": "Get a Mailbox by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseSelected"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Mailbox"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            }
        },
        "/mails/message": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/message"
                ],
                "summary": "Update a Message",
                "parameters": [
                    {
                        "description": "Message Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DepositUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message successfully updated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseUpdated"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Message"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/vcago.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "creates an Message object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/message"
                ],
                "summary": "Create a Message",
                "parameters": [
                    {
                        "description": "Message Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MessageCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Message successfully created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseCreated"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Message"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/vcago.ValidationErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Duplicated Key",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoDuplicatedErrorResponse"
                        }
                    }
                }
            }
        },
        "/mails/message/send_cycular/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/message"
                ],
                "summary": "Get a Message by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message successfully sended",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseSelected"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Message"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            }
        },
        "/mails/message/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/message"
                ],
                "summary": "Get a Message by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message successfully selected",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseSelected"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Message"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/mails/message"
                ],
                "summary": "Delete a Message by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message successfully deleted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseDeleted"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DepositUnitUpdate": {
            "type": "object",
            "properties": {
                "deposit_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "money": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "status": {
                    "type": "string"
                },
                "taking_id": {
                    "type": "string"
                },
                "update_state": {
                    "type": "string"
                }
            }
        },
        "models.DepositUpdate": {
            "type": "object",
            "properties": {
                "crew_id": {
                    "type": "string"
                },
                "date_of_deposit": {
                    "type": "integer"
                },
                "deposit_units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DepositUnitUpdate"
                    }
                },
                "external": {
                    "$ref": "#/definitions/models.External"
                },
                "has_external": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "money": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "status": {
                    "type": "string"
                },
                "update_state": {
                    "type": "string"
                }
            }
        },
...
        "models.External": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "asp": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "organisation": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "reason_for_payment": {
                    "type": "string"
                },
                "reciept": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Mailbox": {
            "type": "object",
            "properties": {
                "draft": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MessageSub"
                    }
                },
                "id": {
                    "type": "string"
                },
                "inbox": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MessageSub"
                    }
                },
                "outbox": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "read": {
                    "type": "boolean"
                },
                "recipient_group": {
                    "$ref": "#/definitions/models.RecipientGroup"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TOData"
                    }
                },
                "to_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.MessageCreate": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "read": {
                    "type": "boolean"
                },
                "recipient_group": {
                    "$ref": "#/definitions/models.RecipientGroup"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "models.MessageSub": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "read": {
                    "type": "boolean"
                },
                "recipient_group": {
                    "$ref": "#/definitions/models.RecipientGroup"
                },
                "subject": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.RecipientGroup": {
            "type": "object",
            "properties": {
                "active_state": {
                    "description": "only used for type crew",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "crew_id": {
                    "description": "only used for type crew",
                    "type": "string"
                },
                "event_id": {
                    "description": "only used for type event",
                    "type": "string"
                },
                "ignore_newsletter": {
                    "type": "boolean"
                },
                "nvm_state": {
                    "description": "only used for type crew",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "state": {
                    "description": "only used for type event",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "can be crew or event",
                    "type": "string"
                }
            }
        },
...
        "models.TOData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "vcago.MongoDuplicatedErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "duplicate key error: \u003ckey\u003e"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "vcago.MongoNoDocumentErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "document not found"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "vcago.ResponseCreated": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_created"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ResponseDeleted": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_deleted"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vcago.ResponseSelected": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_selected"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vcago.ResponseUpdated": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_updated"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "validate error"
                },
                "model": {
                    "type": "string"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "currency is required"
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "vmod.Modified": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "vmod.Money": {
            "description": "Beispiel eines benutzerdefinierten Geldtyps",
            "type": "object",
            "required": [
                "currency"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_cookie",
            "in": "cookie"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://hydra.vivaconagua.org/oauth/authorize",
            "tokenUrl": "https://hydra.vivaconagua.org/oauth/token"
        }
    }
}